refactor(chrome-devtools): split browser tools and storage
